+++ packages/ragbits-document-search/tests/unit/test_web_source.py
@@ -1,12 +1,12 @@
 from sympy.testing import pytest
 
 from ragbits.document_search.documents.exceptions import SourceNotFoundError
-from ragbits.document_search.documents.sources.http import HttpSource
+from ragbits.document_search.documents.sources.web import WebSource
 
 
 def test_id():
-    source = HttpSource(url="http://example.com/file.pdf")
-    expected_id = "https:http://example.com/file.pdf"
+    source = WebSource(uri="http://example.com/file.pdf")
+    expected_id = "web:http://example.com/file.pdf"
     assert source.id == expected_id
 
 
@@ -22,18 +22,18 @@ async def test_from_uri_one_file():
     ]
 
     for uri in file_uris:
-        result = await HttpSource.from_uri(uri)
-        assert result[0] == HttpSource(url=uri)
+        result = await WebSource.from_uri(uri)
+        assert result[0] == WebSource(uri=uri)
 
 
-async def test_invalid_url_raises_exception():
+async def test_invalid_uri_raises_exception():
     wrong_uris = [
         "some string",
-        "https://www.url with spaces.com/path/to/file",
+        "https://www.uri with spaces.com/path/to/file",
         "www.without-protocol.com/path/to/file",
         "http://www.domain-only-with-slash.com/",
     ]
 
     for uri in wrong_uris:
         with pytest.raises(SourceNotFoundError):
-            await HttpSource(url=uri).fetch()
+            await WebSource(uri=uri).fetch()